refactor(economy): phase 3 slice 2 world-market runner consolidation (#3939)
Merge 22 micro-runners into 7 behavior-family files (30 → 11 world_market
tests, meeting ≤12 AC). All 523 tests and description-preservation gates pass;
median wall-clock improved 28.30s → 21.19s.

Refs #3939
