+++ packages/colonizethis_economy/test/economy/world_market/world_market_misc_test.dart
@@ -1,7 +1,7 @@
-// Table-driven SPEC-AC tests for `boycottedColonySellableCommodityIds` (Refs #3856).
-// SPEC/ai/treasury-planner.md § Boycott-aware bid suppression (#3758 S7/R12).
+// Consolidated boycott and lock-recovery runners (Refs #3939 phase 3 slice 2).
 
 import 'package:colonizethis_data/colonizethis_data.dart';
+import 'package:colonizethis_economy/colonizethis_economy.dart';
 import 'package:colonizethis_economy_test_support/colonizethis_economy_test_support.dart';
 import 'package:colonizethis_test/test.dart';
 import 'package:colonizethis_world/colonizethis_world.dart';
@@ -23,6 +23,18 @@ void main() {
           defaultTileMaps: tileMaps,
           defaultTopology: topology,
         );
+      });
+    }
+  });
+
+  group('computeLockRecoveryMinorAutoBids', () {
+    for (final scenario in lockRecoveryMinorBidsScenarios()) {
+      test(scenario.label, () {
+        final bids = computeLockRecoveryMinorAutoBids(
+          game: scenario.game,
+          worldMarketState: lockRecoveryGrainMarket(),
+        );
+        scenario.verify(bids);
       });
     }
   });
