+++ packages/colonizethis_economy/test/economy/world_market/world_market_trade_order_suggester_test.dart
@@ -1,8 +1,10 @@
-// Table-driven unit tests for TradeOrderSuggester.suggest (Refs #3856).
+// Consolidated TradeOrderSuggester and admission runners (Refs #3939 phase 3 slice 2).
 //
-// Treasury-cap (rule 5) cases live in
-// `world_market_trade_order_suggester_treasury_test.dart` (Refs #3123).
+// SPEC/program/world-market-resolution.md § Trade order suggestion API
+// and § Trade order validation.
 
+import 'package:colonizethis_data/colonizethis_data.dart' as data;
+import 'package:colonizethis_economy/colonizethis_economy.dart';
 import 'package:colonizethis_economy_test_support/colonizethis_economy_test_support.dart';
 import 'package:colonizethis_test/test.dart';
 
@@ -51,6 +53,82 @@ void main() {
     for (final scenario in tradeOrderSuggesterValidatorCleanScenarios()) {
       test(scenario.label, () {
         runTradeOrderSuggesterScenario(scenario);
+      });
+    }
+  });
+
+  group('TradeOrderSuggester.suggest — cumulative treasury cap (rule 5)', () {
+    for (final scenario in tradeOrderSuggesterTreasuryCapScenarios()) {
+      test(scenario.label, () {
+        runTradeOrderSuggesterScenario(scenario);
+      });
+    }
+  });
+
+  group('isWorldMarketTradeableCommodity (rule 2)', () {
+    test('non-riches commodity is tradeable', () {
+      expect(isWorldMarketTradeableCommodity('timber'), isTrue);
+      expect(isWorldMarketTradeableCommodity('iron'), isTrue);
+    });
+
+    test('every riches commodity is not tradeable', () {
+      for (final richesId in data.richesCommodityIds) {
+        expect(
+          isWorldMarketTradeableCommodity(richesId),
+          isFalse,
+          reason: '$richesId should be excluded from the world market',
+        );
+      }
+    });
+  });
+
+  group('commoditiesWithBidAndOffer (rule 3)', () {
+    for (final scenario in commoditiesWithBidAndOfferScenarios()) {
+      test(scenario.label, () {
+        final excluded = commoditiesWithBidAndOffer(scenario.proposedOrders);
+        expect(excluded, scenario.expected);
+      });
+    }
+  });
+
+  group('admittedBidCommodityIdsInSubmissionOrder (rule 4)', () {
+    for (final scenario in admittedBidCommodityIdsScenarios()) {
+      test(scenario.label, () {
+        verifyAdmittedBidCommodityIdsScenario(scenario);
+      });
+    }
+  });
+
+  group('offerCapByCommodityId (Refs #3093)', () {
+    for (final scenario in offerCapByCommodityIdScenarios()) {
+      test(scenario.label, () {
+        final game = buildStockpilePlayerGame(stockpile: scenario.stockpile);
+        final cap = offerCapByCommodityId(
+          game: game,
+          playerId: scenario.playerId,
+        );
+        scenario.verify(cap);
+      });
+    }
+  });
+
+  group('stagedOfferQuantitiesByCommodityId (Refs #3093)', () {
+    for (final scenario in stagedOfferQuantitiesByCommodityIdScenarios()) {
+      test(scenario.label, () {
+        final staged = stagedOfferQuantitiesByCommodityId(
+          orders: humanOrdersWith(scenario.orders),
+          playerId: humanPlayerId,
+        );
+        scenario.verify(staged);
+      });
+    }
+  });
+
+  group('sellableHeadroomByCommodityId (Refs #3093)', () {
+    for (final scenario in sellableHeadroomByCommodityIdScenarios()) {
+      test(scenario.label, () {
+        final sellable = runSellableHeadroomScenario(scenario);
+        scenario.verify(sellable);
       });
     }
   });
